requirements.txt
langchain-qwq>=0.3.4
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
